Updates from 03/18/2018 working session and Upgrade to 2017.3.1f1
Tools should now collide with desk and floor.
One of the teleport floors height was changed to prevent it from launching things on the floor into the air. Also the increase in height of the drone "Sphere" (Should probably have better object names) meant that the player could no longer control it when the teleoport floor was on. This was only corrected in the engineering room during this session.