Updated UI to include console feedback
Added a black console like UI to give feedback to the user. May have to add a different way of bringing it up. With this update it will toggle on the spacebar press. It responds to clicks on the system status UI labels to tell the user the state of the system. Original thought is that it would tell the drone how to fix a busted system and they'd have to communicate that to the VR player.