Added "Interactive Mode" to the drone
If the drone player right-clicks the mouse and holds the button down, the player now enters "interactive mode" which will show a cross hair and allow the user to interact with objects in the game. This is primarily to facilitate the drone player "fixing" the broken door at the beginning of the game.

Note: Somebody has to "break" the door