Implemented Power Menu & DEATH CLOCK
The power menu will now show how much power the ship is using. It will show how much is left in the batteries, which at the beginning is pretty much a death clock.

Sure, we don't do anything when we run out of power now, but we will. And when we do, that stupid VR player isn't going to know what hit'em.